Added basic 'parts' framework
Added a base class for special 'Parts' that modify the behaviour of your
battering ram. Going to also try replace the wheels system with just
cylinder colliders and a rigidbody force on the main body - I think it'd
look better to animate the wheels in 2D than have them rotate like this
for pixel art.