Enchantments done
Finished the enchantments I have thought of so far, as well as parts.
Added in a name and description for each in the part database, as well
as adding the correct prefab for each part. A few more parts to go, then
I will likely start working on the UI and/or level stuff.